Move more files to newlines subfolder and clean up related includes
