--- src/do_blank_lines.h
@@ -1,15 +1,0 @@
-/**
- * @file do_blank_lines.h
- * prototype for do_blank_lines.cpp
- *
- * @author  Ben Gardner
- * @author  Guy Maurel
- * extract from newlines.cpp
- * @license GPL v2+
- */
-#ifndef DO_BLANK_LINES_H_INCLUDED
-#define DO_BLANK_LINES_H_INCLUDED
-
-void do_blank_lines();
-
-#endif /* DO_BLANK_LINES_H_INCLUDED */
