+++ src/remove_extra_returns.h
@@ -1,10 +1,8 @@
 /**
  * @file remove_extra_returns.h
- * prototypes for remove_extra_returns.cpp
  *
  * @author  Guy Maurel
  * @license GPL v2+
- * extract from combine.h
  */
 
 #ifndef REMOVE_EXTRA_RETURNS_H_INCLUDED
