+++ src/uncrustify.cpp
@@ -20,7 +20,6 @@
 #include "change_int_types.h"
 #include "compat.h"
 #include "detect.h"
-#include "do_blank_lines.h"
 #include "indent.h"
 #include "keywords.h"
 #include "lang_pawn.h"
@@ -28,6 +27,7 @@
 #include "newlines/add.h"
 #include "newlines/after.h"
 #include "newlines/annotations.h"
+#include "newlines/blank_line.h"
 #include "newlines/chunk_pos.h"
 #include "newlines/class_colon_pos.h"
 #include "newlines/cleanup.h"
